update distro prep to handle staging sample pcaps

+++ distro_prep.sh
@@ -88,24 +88,35 @@ rm -rf ~sansforensics/.config/bless
 /sbin/ifconfig ens33 down
 
 if [ $CASERELOAD -eq 1 ]; then
+    if [ ! -d /mnt/hgfs/sample_pcaps/ -o ! -d /mnt/hgfs/lab_data/ ]; then
+        echo "ERROR: Required source directories in /mnt/hgfs/ are not availalble - exiting."
+        exit 2
+    fi
     labs='demo-01 lab-1.2 lab-2.1 lab-2.2 lab-3.2 lab-3.3 lab-4.1 lab-4.2 lab-5.1 lab-5.2 lab-5.3'
     placeholder_labs='capstone lab-1.1 lab-2.3 lab-3.1 lab-4.3'
     echo "ensure /cases/for572/ only contains what is required from original evidence files"
     echo "will do this automatically, but need to ensure source data is available at /mnt/hgfs/lab_data/ before proceeding"
     read
     rm -rf /cases/for572/*
+
+    mkdir /cases/for572/sample_pcaps/
+    cd /cases/for572/sample_pcaps/
+    cp -a /mnt/hgfs/sample_pcaps/* ./
+
     for lab in $labs; do
         cd /cases/for572/
         mkdir $lab
         cd /cases/for572/$lab/
         unzip /mnt/hgfs/lab_data/${lab}_source_evidence.zip
     done
+
     for placeholder in $placeholder_labs; do
         cd /cases/for572/
         mkdir $placeholder
         cd /cases/for572/$placeholder/
         cp -a /mnt/hgfs/lab_data/placeholders/${placeholder}_readme.txt ./
     done
+
     cd /cases/for572/
     chown -R sansforensics:sansforensics /cases/for572/*
 fi
